requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.1
PyJWT>=2.9.0
Werkzeug>=3.0.0
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Flask-WTF==1.2.1
WTForms==3.0.1
pywebpush==1.14.0
cryptography>=41.0.0
python-dateutil==2.9.0
pytz==2024.1
requests==2.32.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Pillow==10.4.0
